+++ packages/happner-client/RELEASES.md
@@ -359,3 +359,11 @@
 12.3.1 2022-09-01
 -----------------
   - TEN-130: issues with sync init functions and defer startup
+
+12.4.0 2022-10-22
+-----------------
+  - TEN-65, TEN-126: happn-3 updates
+
+12.4.1 2022-10-03
+-----------------
+  - TEN-132, TEN-138: happn and happner updates